requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
bcrypt==4.2.1
email-validator==2.1.0 